requirements-base.txt
django==2.2.12
django-environ==0.4.5
django-filter==2.2.0
django-localized-fields==5.4.2
django_mptt==0.11.0
django-postgres-extra==1.22
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
uwsgi==2.0.18
